+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Changelog
 
+### Next Release
+
+ - Deprecated `arguments:` in favor of `_args:`
+
 ### `0.7.0`
 
  - Introduce new `--switch` syntax (`switch` without a following colon) (#30)